refactor(tests): move VCR configuration to tests/providers
